+++ src/main/java/com/ryc/api/v1/role/dto/response/GetClubRoleApplicationResponse.java
@@ -1,4 +1,4 @@
-package com.ryc.api.v1.role.dto;
+package com.ryc.api.v1.role.dto.response;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
